Fixed a bug in the continuum mode

+++ ptsmodel/_ptsmodel.py
@@ -881,9 +881,15 @@ class PTSMODEL():
         '''
         # files
         path_infiles = model_utils.__file__.split('ptsmodel')[0]+'infiles/'
+        fins = ['dustkappa_%s.inp'%self.dustopac, 'molecule_%s.inp'%self.line] if self.line is not None\
+        else = [ 'dustkappa_%s.inp'%self.dustopac ]
         for fin in ['dustkappa_%s.inp'%self.dustopac, 'molecule_%s.inp'%self.line]:
             if os.path.exists(fin) == False:
-                shutil.copy(path_infiles + fin, '.')
+                if os.path.exists(path_infiles + fin):
+                    shutil.copy(path_infiles + fin, '.')
+                else:
+                    print('WARNING: Cannot find necessary files.')
+                    print('WARNING: File name --- %s.'%fin)
 
         print ('radmc3d mctherm setthreads %i'%nthreads)
         os.system('radmc3d mctherm setthreads %i'%nthreads)
